Network/Util: Extract network util functions from MainService
This patch extracts network realted util functions from MainService.
